requirements.txt
tweepy
kafka-python
pyspark